Fix some bugs in VSSessionTranslator
Summary:
This fixes a couple misc bugs due to races when the backend of a VS Code debug protocol engine is totally asnychronous. Some of the messages are arriving to the UX out of order due to flatMap instead of concatMap, and stack trace requests can be fired off due to the initial loader break, and then if the target resumes before the stack trace commands return, and the backend responds with a failure for the stack trace commands, the UX layer breaks and fails to update anything.

Also this fixes a bug where the UX still shows an old hitcount for a breakpoint after disabling it.

Reviewed By: mostafaeweda

Differential Revision: D6530786

fbshipit-source-id: c1d1cff13cc1f534b187cff70b6fde5d96fb639a