Add vscode-uri to handle file:// protocol
Summary: Currently our file:// protocol conversion is a little broken for special character path escaping. The paths are sent uri-encoded meaning that '#','?',' ' and similar characters are percent-escaped, but we do not properly decode on receiving and encode on sending (e.g. '%' and '/' are double-escaped). We can use the vscode-uri package to provide us with the same uri handling our LSP providers would use.

Reviewed By: hansonw

Differential Revision: D6540203

fbshipit-source-id: f84c6903034ae35b0ce04c6e08a7236e2ea0c94b