fix device visibility
Summary:
If the list of devices is not changing and the action setDeviceType is triggered, without actually changing the device type, which could happen due to some useful logic inside the react components, then the pending state is shown until the list of devices changes.
That happens because the device observer only emits when the list of devices changes.

This fixes this issue by not modifying the state when setDeviceType is called without actual changes.

Reviewed By: stepanhruda

Differential Revision: D6558094

fbshipit-source-id: 144a2a918e6eb407f59fac1349ec9e2974327855