Add new 'fromFilesystem' property when fetching RevisionInfo
Summary: Currently there is no way to figure out if the contents came from the initial empty value or were actually fetched. The issue that is being solved here is that the initial fetch for revision status is fetched from a cache that has an initial value of an empty array! This change adds a new fromFilesystem property that will let the subscribers know if the value was from the default start value or the file system.

Reviewed By: evangrayk

Differential Revision: D6548641

fbshipit-source-id: 5e3ee824a36fa5b934f3c9ae591b21791e19ac45