fix adapter launch on dev server
Summary:
On the dev server, the debug adapter is being launched with
the wrong version of node. Pass in the correct node for launching
adapters.

Also, ensure that the new version of node is used for launching
the debug target, as the devserver version speaks the old version
of the protocol and can't be attached to.

Reviewed By: ebluestein

Differential Revision: D6561009

fbshipit-source-id: e0ae7639047b1338ab896da2f322fb41485b514d