Prompt to file a bug if the local RPC process crashed
Summary:
We track these as fatals in logview, but it's good to have error reports too.
Note that we also have to error out all `send()` calls after a crash or else the bug report hangs gathering logs. And since this is open source we have to verify the presence of fb-file-a-bug.

Reviewed By: mostafaeweda

Differential Revision: D6564938

fbshipit-source-id: a07967169f1ac809002abdc98ca42986ebaf0a5b