Fix potential NPE in getOpenFileEditorForRemoteProject
Summary:
Right now we're checking one thing (getURI) and using another (getPath).

In both instances (`TextEditor` and `RemoteTextEditorPlaceholder`) the nullability of the two should be consistent... but ljw1004 has a clear report where this wasn't the case.
I have no idea how this happened but it seems like the right thing to do regardless!

Reviewed By: mostafaeweda

Differential Revision: D6569929

fbshipit-source-id: 317b9f32f92060592192be2b4a8d6e19d7e7e3dc