JS imports: Resolve files relative to yarn workspaces too
Summary:
Nuclide's `modules/` Yarn workspace is not part of the .flowconfig anymore.
To resolve these correctly, the JS imports server needs to know about Yarn workspaces too (which are declared as `workspaces` in `package.json`).

Due to a lack of foresight I neglected to include `moduleDirs` in the cachekey; this is now remedied.

Reviewed By: wanderley

Differential Revision: D6540254

fbshipit-source-id: 02549c8202a3e41678042d021a6d108ccf541530