Surface errors/warnings from fetching buck compile commands
Summary: Previously we caught all errors and returned null in their place. Now we pop up a notification box and log the relevant portion in the console window. Also add a warnings key to the Buck compilation database object to track issues (e.g. missing owners) which that we attempted to work around.

Reviewed By: stepanhruda

Differential Revision: D6574137

fbshipit-source-id: ffcd74bd6286548000ca3ed138874267b70cab02