Fix memory leak with TerminalView
Summary:
Right now all TerminalViews are retained forever because the server-side `PtyImplementation` never frees references to the `PtyClient`.

This is a super subtle behavior - if we don't dispose `PtyClient` on the server, the client must keep a reference around since it doesn't know if the server is going to call something on `PtyClient` again.

This flow is subtle because it actually triggers a double-disposal. Looking at the `dispose()` implementation in `TerminalView` however it appears that this is quite safe to do.

Reviewed By: mostafaeweda

Differential Revision: D6582084

fbshipit-source-id: 5fcde5ce787a32b0166b0711fdb3ff201b058aea