Mounted component test: replace use of React internals with patching of ReactDOM's API
Summary:
This replaces our use of React internals for accessing unmounted root components with a simple monkeypatching of the core ReactDOM API.

A map of mounted React elements is maintained, and elements are removed from this map right before they're unmounted from the dom.

This will remove some tech debt and allow us to update React to 16.2, which entirely removes the introspection we previously used.

Reviewed By: matthewwithanm

Differential Revision: D6598511

fbshipit-source-id: 59dc001d6819a74c35cab78dd7b12ccf4ad9a063