Move text edits sort to inside applyTextEditsToBuffer
Summary: I found a call-site of `applySortedTextEditsToBuffer` that was not sorting the `textEdits` before sending. After chat with hansonw, I got to a conclusion that is better to not let to caller the responsibility to sort the `textEdits`. This diff pushes to `apply*TextEdits*` functions the responsibility to sort the input.

Reviewed By: hansonw

Differential Revision: D6573857

fbshipit-source-id: b7a3478fe3ef62584937c36e9b17b9d8dfe2eeea