Create nuclide-autocomplete package
Summary: Autocomplete feature is really important for our users. This diff creates a proxy package that should be used when providing a new autocomplete functionality to Nuclide. This package will take care to track timing for autocomplete operations and in a future diff it will enable introduce a timeout for autocomplete functions.

Reviewed By: hansonw

Differential Revision: D6574118

fbshipit-source-id: 90646a34b54cb5e2de76418a53cbff4e3fd892e7