Use nuclide-autocomplete on nuclide-language-service
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: hansonw

Differential Revision: D6574923

fbshipit-source-id: 7dc1605f8cc51fc170e3461b9ef8ed296358d5d2