Use nuclide-autocomplete on nuclide-clang
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: a20012251

Differential Revision: D6574920

fbshipit-source-id: ee144a8eff7c1144a6776593997897252f34001d