Use nuclide-autocomplete on nuclide-console
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: pelmers

Differential Revision: D6574927

fbshipit-source-id: cfd948c614694ace0310eaf3cba5b946a837d162