Use nuclide-autocomplete on nuclide-debugger
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: mostafaeweda

Differential Revision: D6574929

fbshipit-source-id: 9c27dac0480366a0533877562b36b553a4bf20fa