Use nuclide-autocomplete on nuclide-swift
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: modocache

Differential Revision: D6574926

fbshipit-source-id: e03f887983f06354bf9023e17cb1ef4b89113d14