Add toggle button for blame gutter
Summary:
This diff adds a button in the lower right corner of an editor that toggles the blame gutter for an editor, if the editor is capable of displaying it.

The implementation actually adds the button to every editor, and hides / shows it based on the availability of a BlameProvider for the editor, and the editor's state.  Visibility updates are triggered on changes to the editor buffer (`atom$TextEditor.onDidStopChanging`), save (`atom$TextEditor.onSave`), and on initial package load (`atom.packages.onDidActivateInitialPackages`).  The latter is needed because buttons are added on editor creation, but editors may be created before all packages, including blame providers, are loaded.

Reviewed By: aadisriram

Differential Revision: D6449215

fbshipit-source-id: 646fc6d546e43e28fd3f542a9cb65f8372d87d7e