Use Hack fragment grammar consistently where appropriate
Summary:
I created a custom Hack fragment grammar a long time ago because normally the Hack/PHP grammars require a '<?php' or '<?hh' block to start highlighting anything.

This creates a common `getFragmentGrammar` helper to accomplish this (if `source.hackfragment` is available). Note that this isn't the case for most atom-ide-ui users, but it's hard to pull out the dependency from find-references.

Reviewed By: tjfryan

Differential Revision: D6598311

fbshipit-source-id: fdafcd54c8e25aef6a94f8c4d9ca91ac319c165c