Revert HostAggregator.syncProgress
Summary:
We introduced syncProgress as a temporary stop-gap because nuclide-rpc sometimes lost messages from server to client.

Now that nuclide-rpc has been made robust, we can ditch it.

(There was also a report of a use-after-dispose bug within syncProgress. It feels easier to delete the code than to first debug it and then delete the code...)

Reviewed By: hansonw

Differential Revision: D6618645

fbshipit-source-id: d543d6218ae44002f4e6629334c42e11535e6102