HostServices.showProgress race
Summary:
There was a report of a "use-after-dispose" error in HostAggregator.js. The error was in UniversalDisposable.add, called from HostServicesAggregator.js. Here's what I think happened:

This fix works by making sure that in the final step, it will follow the "cancelled" path, which doesn't call relay._disposables.add()

Note: While writing the unit tests, I noticed that a bunch of mocks were returning objects rather than Promise<objects> as they should. This so far had no effect because we were always awaiting the result; MDN explains "If the value of the expression following the await operator is not a Promise, it's converted to a resolved Promise.". I noticed the issue now because I exercised a path that called .then() on the result.

Reviewed By: hansonw

Differential Revision: D6618964

fbshipit-source-id: 917fbee9c5a7793d085b65b64b7f9bb6cfef9e51