Use nuclide-autocomple on nuclide-ocaml
Summary:
The `nuclide-autocomplete` package is a proxy package for `autocomplete-plus`. It makes sure that an autocomplete provider has `getSuggestion` and `onDidInsertSuggestion` times tracked. In future, it will have an expection handling logic and a timeout to make sure that slow autocomplete providers hold the autocomplete response.

This diff updates this package to use `nuclide-autocomplete` instead of `autocomplete-plus`.

Reviewed By: hansonw

Differential Revision: D6574925

fbshipit-source-id: 275ffcf54992dda51160f52b4b86c4a149d6259e