Automatically set the device sandbox hostname when the FB App starts
Summary:
This diff uses the last two services added -- 1) get a list of the user's On Demand instances, and 2) know when a buck task has finished -- to ask the user if they want to update the sandbox hostname of the FB App to their last On Demand WWW instance. This is triggered when the user installed a new app on the device and when its sandbox is different than the On Demand WWW.

This change is behind a GK `nuclide_on_demand_set_sandbox_on_launch_app`, and this feature will be test with a selected number of users.

I also created some helper functions to implement this:
* **getRemoteDevice**: given a device updates its ports to use the remote adb
* **tryToSetSandboxOnDevice**: tries to set the sandbox by reading the sandbox (to make sure they're different) and retrying a specific amount of times before giving up. It needs to do this because we don't know how long it will take for the just launched app to be up and running.
* **getSandboxOnDevice**: Does what it says and works with remote devices.

Reviewed By: hansonw

Differential Revision: D6639696

fbshipit-source-id: 7f71503d6b423045b4c550a1372489868a3d1133