HHVM VSDEbugExtension - Support Launch/Attach protocol in wrapper
Summary:
This change updates the HHVM debugger wrapper to support attach as well as launch, and to connect to the target in response to a LaunchRequest or AttachRequest VS Code Debugger protocol command, as opposed to a command line arg.

The wrapper now handles the proper attach flow:

1. The debugger client sends an InitializeRequest. The wrapper has no use for this since any logic that depends on it is going to be in HHVM. The wrapper saves this to forward to HHVM (which we haven't connected to yet).

2. The debugger client sends a LaunchRequest or AttachRequest - the wrapper then either launches HHVM or attaches to it via TCP.

3. The wrapper forwards the InitializeRequest to the debugger extension within HHVM, which can then do whatever it needs to with it.

It may be necessary for the wrapper to look at the path format in the initialize request if we support non-default paths (URI vs native path format). Nuclide is going to use native path format though.

Reviewed By: rightparen

Differential Revision: D6266330

fbshipit-source-id: 33d7ec685f29a5d7fb22a175bc960dab0111787b