Nuclide HHVM Debugger - Wire up Nuclide to new debugger!
Summary:
This change adds a new HhvmDebuggerService that will replace the existing PhpDebuggerService eventually (leaving the old one in place until we're ready to switch).

On the front-end a GK is checked and if it passes, we instantiate the new debugger instead of the old one. The launch/attach args are majorly simplified now and cut out all the XDebug nonsense (idekey, xdebug port, log levels, ...).

The only logic saved/copied from the old debugger is:
  - Finding hack root to resolve startup document
  - Finding path to Node.js

We currently have a feature config setting in the PHP package for path to PHP and args for PHP runtime. I am copying those to new settings that start with 'hhvm'. It is confusing to have both settings in there, but I'll need to leave it like this for a little while, just until the new debugger is ready and we drop support for the old one. The rename is a good idea because a.) we're not going to support PHP anymore. HHVM will be the only PHP-ish runtime we're going to debug from Nuclide.   b.) Any users with an existing PHP path in there, we don't want to use that if it's pointing to PHP

Reviewed By: rightparen

Differential Revision: D6274504

fbshipit-source-id: 6b43fc4af9201182ca2b54a111e258689254ee5b