Lazily request scope variables from UX
Summary:
Currently our LazyNestedValueComponent correctly only requests a scope object's properties when it is expanded. But the top level scopes (Locals, Globals, etc) are all requested right away, even if the scope is collapsed and this is done for every thread.

This is super wasteful and expensive for things like Globals in PHP (there are 100s of them). We should follow the same behavior that we do for the lower levels like properties on objects, and only request if the user expands the scope section.

Reviewed By: mostafaeweda

Differential Revision: D6604034

fbshipit-source-id: 489f12bc01eb47a02e1adecfbf437070555e6c82