Upload Flow logs in flytrap
Summary:
When users encounter errors in Nuclide, they use the bugnub to
report them. Sometimes these users are reporting Flow errors, and these
tasks get assigned to the Flow oncall. However, Flow logs are not
uploaded, so the oncall is out of luck.

This diff attempts to upload the server logs (`.log` and `.log.old`),
the monitor logs (`.monitor_log` and `.monitor_log.old`) and the pids
file (`.pids`). If any file is above 10MB, we only upload the last 10MB
of the file.

Reviewed By: hansonw

Differential Revision: D6641002

fbshipit-source-id: c2989ac7d6a870610365725d4c1ec1646cfa798f