Send different empty components
Summary:
In the device table list, if the contents of an empty component were supposed to change, the node wouldn't get rerendered. From the `Table`'s point of view, the empty component was always represented by the same prop, which it runs shallowequal on, even if the arrow function would yield a different result.

The easiest way to reproduce this was to:

1. Open Nuclide with no Android devices available
2. Check device panel, the table had a spinner showing indefinitely.
3. The spinner would switch to the correct `No devices found` node if you forced the table to re-render by eg switching to iOS and back (providing iOS had any rows available to trigger the re-render).

Reviewed By: a20012251

Differential Revision: D6714485

fbshipit-source-id: 7e8bf72cd07bb3e3630b86a48f0cd162a6bbddf3