Nuclide HHVM Debugger - Implement log flushing and rolling - 23/N
Summary:
This diff introduces a change for logging whereby when a debugger client is attached, we periodically flush the log file (every 1 min), and rotate the logs (every ~24 hours) saving the most recent 7 log files only. This prevents the log file from growing indefinitely if logging is enabled.

We only do this work while a client is attached, so we don't pay for flushing the log every 1 min all the time.

Also note, logging is disabled by default, unless explicitly turned on in config.hdf.

Reviewed By: rightparen

Differential Revision: D6599280

fbshipit-source-id: 47cdc0f7e29962285ac7967ce727acd966cfc3d2