Nuclide HHVM Debugger - Allow configuring log file path in launch/attach command - 34/N
Summary:
This allows the client to specify a log file path to use when launching or attaching, and updates Nuclide to include the HHVM debugger logs when filing a bug via flytrap.

This will allow us to configure logs from the Nuclide side, without having to push an update to HHVM config to it, and to know where the logs are stored so that we can grab the when filing bugs.

Rather than streaming the log down to the Nuclide client, the log is sent as a Paste, and the client logs a file containing just the paste URL.

Reviewed By: rightparen

Differential Revision: D6669891

fbshipit-source-id: 1128c6c470dc1b728a5dcb6b6bd7658a0dc3bafe