Nuclide HHVM Debugger - Fix races in prepare to pause logic - 38/N
Summary:
This fixes a race between calling prepareToPauseTarget, which ensures the target is in a consistent state, and actually entering the command queue because the existing code drops m_lock before calling processCommandQueue.  processCommandQueue then re-acquires the lock briefly, updates the number of paused threads and the program state, and then drops it before entering the queue.

The correct thing to do is hold the lock the entire time, and drop it just before entering the command queue, after the program state and the number of paused threads have been consistently set under the lock.

This also cleans up the logic in prepareToPauseTarget and fixes a potential (though very unlikely) timing bug where an in-progress async-break could spin forever if another thread resumes and then quickly pauses the target.

Reviewed By: rightparen

Differential Revision: D6690565

fbshipit-source-id: add7706a6de2bf7719283cffce2c4b3ce41649e4