Add config for Tizen
Summary:
Tizen is an obscure platform compared to Android and iOS. Add a checkbox to optionally hide it.

Implementing a service registration/deregistration seems like an overkill, seems okay to need a restart for now.

This could be better by eg having nuclide-tizen, but seems cheap and good enough for now.

Reviewed By: a20012251

Differential Revision: D6730931

fbshipit-source-id: 178a6157a30c969106bd8ac4ff11b3dc373c6b6f