Add icon to remove breakpoint
Summary:
* Adds a new "X" icon next to the edit icon that will delete a breakpoint.
* Makes a few minor layout changes to account for the new icon, including vertical centering
* Adds a new Analytics event for deleting a breakpoint from the icon. I'm not sure if this is desired, but it's consistent with the edit icon

I looked over the tests to see if I could add a test for this, but I'm not exactly sure how to test items in the Debugger pane. There wasn't a test (I could find) for the edit icon, so any info on how to write a test for this would be great.

Reviewed By: ebluestein

Differential Revision: D6732482

fbshipit-source-id: 4315194b912375533d2c18d5a419d295a36f4f48