set hypertext link handler in terminal
Summary: xterm.js 2.x.x added anchor elements into the document representing the shell session. With canvas in xterm.js 3.x.x, the default is to call window.open(), which doesn't do anything in Atom. This adds a handler to forward the request to shell.openExternal(), which opens the link in the default browser.

Reviewed By: pelmers

Differential Revision: D6763208

fbshipit-source-id: 0b7e24a30b8a4ba2404c3e301705e06f5e0a80ff