Add observeTunnels to service
Summary:
I was planning on using this in another package. I am not 100% sure I will, but I already wrote it and it still seems as a useful API for a consumer.

It can replace getOpenTunnels() since it sends the current value on subscription, but I didn't bother removing that function yet.

Reviewed By: semmypurewal

Differential Revision: D6763946

fbshipit-source-id: 2151b9528dc12d3f5e4138dcb2fd3cc03d0987b2