Send sys_path for each language service request
Summary:
Jedi accepts a `sys_path` that acts as the module resolution path(s) for Python imports.
We used to call `add_paths` after link trees and top level module paths were resolved.. but the main drawback here is that if the server crashes and restarts then we've lost all the additional paths :(

In addition if we want to merge all the servers together (the ultimate goal) the Python server would have to keep track of all the paths for each of the files, which is kinda annoying.

Since there usually aren't that many `sys_path` entries, it's easiest to just send them along with each language server request.

Reviewed By: pelmers

Differential Revision: D6764599

fbshipit-source-id: 4832b52d546df897767cf858c8701ba32fd2c85b