Fix Nuclide bug around tag @generated
Summary:
Cause:
Currently Nuclide will cache a file's status to readonly when the file contains 'generated' this cache is not cleared when user eliminates the string 'generated'. Thus, the file is always marked as readonly, even after user replaced 'generated', saved and reopen the file.

Fix:
On file save, clear the cache for generated tag.

Reviewed By: drmathochist

Differential Revision: D6776062

fbshipit-source-id: d9c238c15186372cae6c8a92f88747711740269e