Fix terminal resizing
Summary: The 'fit' add-on of terminal is not expecting to be used with a div whose width is '100%'. When such a div is display:none, its computed style has a width of '100%' rather than a pixel count. The result is that things inside the terminal get squished to fit in 100 pixels. It also turns out that we don't need this--the div takes up the full space by default. With nothing specified and display:none, the widht/height is NaN and ignored.

Reviewed By: pelmers

Differential Revision: D6792928

fbshipit-source-id: bf80dc188f68422c771ca3c40eadcc15f8cf253d