Basic artillery tracing for nuclide
Summary:
This change adds a very simple fb-internal tracing library to nuclide. Callers of the library simply call NuclideArtilleryTrace.begin() to mark the start of the event they want to trace, and .end() to finish it. For users outside fb, these are no-ops.

As an initial demonstration / use case, this diff also adds tracing for buck commands.

Reviewed By: hansonw

Differential Revision: D6776668

fbshipit-source-id: eb6f4d6f33ed3a477453318c850800804e93a526