Revamp link-tree resolution
Summary:
The current `rdeps` logic is inefficient for two reasons:

1) it has separate queries for binary/unittest targets, but we can combine them
2) it issues queries for directories that don't have target files, which is a waste of time

Instead we'll switch to a two-step process that 1) resolves the owner of the file and then 2) looks for both binary+unittest rdeps among its immediate neighbors.

This also memoizes all the intermediate (and final) computations. In a future diff, I'll likely be re-using all of the intermediate results to prompt the user to build something that has link trees if possible.

Finally, while I'm here, I disabled link-tree lookups for BUCK/TARGETS files since they're treated as Python but obviously don't have linktrees.

Reviewed By: pelmers

Differential Revision: D6781656

fbshipit-source-id: 3d338caac09bcf075b5819648250f7f4e82eabfc