Add registration of arbitrary components
Summary:
Adding new features to the device panel pretty much always requires a new type of provider and changes to the core package itself to display the new data.

This allows consumers to pass an arbitrary UI component instead. The component gets the list of devices as props as a courtesy. We could add a similar provider for the individual device page.

This should allow us to prevent overfitting the service UI to a use case applicable to a single device type, e.g. the process table won't work on iOS since we don't have access to processes.

Reviewed By: a20012251

Differential Revision: D6798927

fbshipit-source-id: e7bdd75201fb52962f2ae1d42eefc322c439f8c7