Expose highlighted tree nodes as a separate structure
Summary:
Rather than modifying properties of the outline structure and rerendering the outline with a new outline object on selection changes, keep that tree constant while updating a list of paths to highlighted nodes.

This adds support for multiple selections in the tree, which is not currently used. Multiple cursors would be a good application of this.

This appears to preserve the perf optimizations hansonw made a while back, though in a different way: `distinctUntilChanged` prevents the component from rerendering unless either the outline or highlighted paths have actually changed.

Externalizing these paths paves the way to more easily implementing keyboard navigation in Outline.

Reviewed By: matthewwithanm

Differential Revision: D6760918

fbshipit-source-id: 5a748b4a10bcc8303d74345d46c6c0a7ec6bd47b