Fork <Tree> to <SelectableTree>
Summary:
This is so I can freely break all the API contracts that exist without risking breaking other trees in the system.

This also only exports a single `<Tree>` component, which is far more than presentational like the current Tree.js. It accepts a prop, "items", which describes how the nodes in the tree should appear rather than accepting any type of children.

This type of tree accepts a list of selectedNodes and a handler for when a node is selected.

This tree provides the basis for keyboard navigation.

Reviewed By: matthewwithanm

Differential Revision: D6778867

fbshipit-source-id: ee1aba1f0029bfa0028e3f721ef87bb54eb787c4