Add in subscribeToFileCreateAndDelete
Summary:
We want to be able to observe a file being created&deleted by hg easily. This will allow us to follow `.hg/wlock`, as well as other status files that indicate merge in progress, rebase in progress, etc.

This change adds an abstraction that gives you a promise for a watchman subscription where you can provide a callback for each emitted value. This way it fits with other code in hgService so you can await multiple such promises together.

The creation/deletion watching is done with a normal watchman subscription in which we check the `new` and `exists` fields to determine if the file was created or delete in this event. We also do some filtering to eliminate events where a file was updated but not created or destroyed, since we don't care about those events right now.

Watchman does some debouncing of events, but they get coalesced in a way that we can rely on it providing us create/delete events. The added tests cover cases where there is a delay between events and cases where there are lots of events in a short time.

This function actually allows you watch for multiple files at once, to prevent too many watchmanSubscriptions. The provided callback is called with a map of file name -> exists as a boolean, so that the consumer can see what has changed and keep track.

Reviewed By: tjfryan

Differential Revision: D6778939

fbshipit-source-id: bf80f8042220a55f01b384608317d6de03ecd45a