Add hg lockFiles observable to HgService
Summary:
Use previous diff to subscribe to the creation and deletion of the `.hg/wlock` and `.hg/histedit-state` files, and expose this as an observable in HgRepositoryClient/HgService. This will be used to 1. prevent the user from saving the file while the workign dir lock is active and 2. display an indicator that hg is working on something in the status bar at the bottom.

I'm adding two files at once because I needed to test that my code actually works for multiple files, so hopefully it isn't too much overhead to review both at once.

Reviewed By: tjfryan

Differential Revision: D6779377

fbshipit-source-id: 36d88c19caf6691ad65a11874cf4a6c691fcd003