Display code actions for link tree targets
Summary:
If we detect that a file has usable linktrees but none of them have actually been built (if they don't exist), we can display a diagnostic with code actions to open the Buck toolbar.

Note that this only works when the CWD is a valid Buck root, so I had to add a qualifying check for that.

Reviewed By: pelmers

Differential Revision: D6799439

fbshipit-source-id: 4760a23e028e975e9a75c6ba1752213c93784232