ObservablePool: a generalization of PromisePool
Summary:
`nuclide-commons/promise-executors` currently has a `PromisePool`, which
allows you to run promise-returning functions with a concurrency limit.

This essentially creates an analogous `ObservablePool` that allows you to
execute functions that return any Observable input
(i.e. Observables, Promises, or Iterables) with a concurrency limit.

The additional power of `ObservablePool` is its ability to
forward unsubscriptions, even if execution is still queued (if a request
is cancelled while in the queue, it never executes).

For requests that return a Promise, the ObservablePool is pessimistic
and assumes that the operation is uncancellable - it will not remove
the execution from the pool until it resolves or rejects. However
`execute()` will still return an Observable to enable the use case
of cancelling requests while they're in the queue.

Reviewed By: matthewwithanm

Differential Revision: D6829550

fbshipit-source-id: 9a67b9bd9832093ab029b393fb8b894f96d0f031