Update Nuclide language service find-references return type
Summary:
The main reason for this change is that the remote service framework has a 60 second timeout for promises. I have seen find-refs requests take longer than that. Since find-refs will often be a long-running command, I think it's reasonable to lift this restriction.

This change may also be nice if we want to get set up to stream find-refs requests in as they are available, or to provide cancellation. `Observable`s allow both of these behaviors, though some much more pervasive changes would be necessary to accomplish them.

Reviewed By: drmathochist

Differential Revision: D6823924

fbshipit-source-id: e36055a8770296d88f5496122259d0028aa09be2