Prevent DataTips from triggering from the gutter.
Summary:
`text.pixelPositionForMouseEvent` clamps it's values to the dimensions of the scrollable area, so any event to the left of the editor will be reported as a value of exactly zero.

This meant that if you had a diagnostic in column 0 (as you do with Sample Diagnostics Tester), and you hovered over the gutter item associated with it, it would show both the gutter popover _and_ the text datatip.

We lose responsiveness in the far left pixel of the editor, but I think that's acceptable.

Reviewed By: matthewwithanm

Differential Revision: D6850631

fbshipit-source-id: d26418d453f0f4ee7be39ffce9671cf36a57e24d