Pool together references to clang flags
Summary: Instead of storing a flags array for each file, pool the arrays and just store an index instead. In the best case we save lots of memory, and in the worst case we are the same as before (at most one extra integer per file). Even large databases will only have a few thousand files so linear searching the flag pool should not be a big concern.

Reviewed By: a20012251

Differential Revision: D6841682

fbshipit-source-id: 17387667b1b7b8e7e485d223cd12b146041dee38