Create the big-dig-based Nuclide server (nuclide-server2)
Summary:
With the big-dig library in place, it's relatively easy to spin up a Nuclide server!
(Although, there are still a lot of todos to bring this to parity).

I decided to hijack the existing `nuclide-start-server` script by checking for the presence of a `--big-dig` flag (which we'll invisibly append to your remote server command if you pass the GK).

Reviewed By: semmypurewal

Differential Revision: D6809750

fbshipit-source-id: 77775af1860d62f3fb0feaea28838dc47c26c685