Allow RemoteConnection/ServerConnection to work with big-dig servers
Summary:
This adds a `version` parameter to the `RemoteConnection` / `ServerConnection` configs. If `version = 2` is supplied, the underlying `ServerConnection` will create a `BigDigClient` rather than the usual `NuclideSocket` to connect to the server.

Not much refactoring was necessary to get this to work. However, the server heartbeat logic is still tied to the `NuclideSocket` implementation so I've just stubbed those out for now. (wanderley is working on T25532978 to make that abstract.)

Reviewed By: semmypurewal

Differential Revision: D6809738

fbshipit-source-id: ea75c9a56b441c3dae51001579e60f6ebaff0612